+++ independent-projects/panache/src/main/java/org/jboss/panache/controller/Controller.java
@@ -1,4 +1,4 @@
-package org.jboss.panache.jpa;
+package org.jboss.panache.controller;
 
 import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.Response;
